implement Weighted Derivative DTW

+++ src/algorithms/lcss.cpp
@@ -16,17 +16,18 @@ bool LCSS::match(Quaternion *vec1, Quaternion *vec2, float epsilon,
 
 #define INDEX(x, y) ((x)* (m + 1) + (y))
 
-float  LCSS::lcss(const std::vector<Quaternion *> &inpF, const std::vector<Quaternion *> &refF, float epsilon, float delta,
-                  std::function<float(const Quaternion *, const Quaternion *)> func) {
+float
+LCSS::lcss(const std::vector<Quaternion *> &inpF, const std::vector<Quaternion *> &refF, float epsilon, float delta,
+           std::function<float(const Quaternion *, const Quaternion *)> func) {
     size_t n = inpF.size();
     size_t m = refF.size();
-    float* L = new float[(n + 1) * (m + 1)]();
+    float *L = new float[(n + 1) * (m + 1)]();
     for (int i = 1; i <= n; i++) {
         for (int j = 1; j <= m; j++) {
-            if (std::abs(i - j) <= delta && func(inpF[i - 1], refF[j - 1]) < epsilon) {
+            if (static_cast<float>(std::abs(i - j)) <= delta && func(inpF[i - 1], refF[j - 1]) < epsilon) {
                 L[INDEX(i, j)] = 1 + L[INDEX(i - 1, j - 1)];
             } else {
-                L[INDEX(i, j)] = std::max(L[INDEX(i - 1,j)], L[INDEX(i, j - 1)]);
+                L[INDEX(i, j)] = std::max(L[INDEX(i - 1, j)], L[INDEX(i, j - 1)]);
             }
         }
     }
